backend/requirements.txt
# FastAPI Backend for RAG Chatbot
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.1
pydantic>=2.9.0

# Qdrant Vector Database
qdrant-client>=1.12.0

# LangChain for RAG
langchain>=0.3.0
langchain-community>=0.3.0
langchain-google-genai>=2.0.0
langchain-qdrant>=0.2.0

# Text Processing
markdown>=3.7
beautifulsoup4>=4.12.0
lxml>=5.3.0

# Embeddings (multilingual for Urdu support)
sentence-transformers>=3.2.0

# CORS support
fastapi-cors>=0.0.6

# NumPy (use pre-built wheels)
numpy>=1.26.0
